# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Visit Lab",
        "contact": {
            "email": "saviogmoiagaia@gmail.com"
        },
        "version": "3.0.0"
    },
    "paths": {
...
        "/api/visits": {
            "get": {
                "tags": [
                    "Visitas"
                ],
                "summary": "Retorna as visitas dos laboratorios",
                "operationId": "bc48af561c011aaa53460ba7418472c0",
                "responses": {
                    "200": {
                        "description": "Retorna as visitas dos laboratorios"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/api/visit": {
            "post": {
                "tags": [
                    "Visitas"
                ],
                "summary": "Cria uma visita.",
                "description": "Cria uma visita",
                "operationId": "dd0e97173da710be6a2586eabe8caa3a",
                "requestBody": {
                    "description": "Json com informações necessárias",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "visitor_name",
                                    "visitor_document"
                                ],
                                "properties": {
                                    "visitor_name": {
                                        "description": "Nome do visitante",
                                        "type": "string"
                                    },
                                    "visitor_document": {
                                        "description": "Identificação do visitante",
                                        "type": "string"
                                    },
                                    "laboratory_id": {
                                        "description": "id do laboratorio onde foi feita visita",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Formulário inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/api/visits/{visit}": {
            "get": {
                "tags": [
                    "Visitas"
                ],
                "summary": "Obtém uma visita pelo id",
                "operationId": "cb8723179b9e5dd335f103ac055eacc4",
                "parameters": [
                    {
                        "name": "visit",
                        "in": "path",
                        "description": "Id da visita",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pega um laboratorio"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Visitas"
                ],
                "summary": "Atualiza uma visita.",
                "description": "Atualiza uma visita, colocando a data e hora de saída",
                "operationId": "099f41a7ce3ce333c3f1cde29de8ba05",
                "parameters": [
                    {
                        "name": "visit",
                        "in": "path",
                        "description": "Id da visita",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Formulário inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    },
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerToken": {
                "type": "http",
                "description": "Access token obtido na autenticação",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "api_key": {
                "type": "apiKey",
                "description": "Chave de aplicação para autenticação na API",
                "name": "X-Application-Key",
                "in": "header"
            }
        }
    }
}